requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
asgiref==3.7.2
astroid==2.15.6
async-timeout==4.0.3
asyncio==3.4.3
asyncpg==0.28.0
attrs==23.1.0
backports.zoneinfo==0.2.1
black==23.7.0
charset-normalizer==3.2.0
click==8.1.7
dill==0.3.7
Django==4.2.4
django-filter==23.2
djangorestframework==3.14.0
frozenlist==1.4.0
greenlet==2.0.2
idna==3.4
isort==5.12.0
lazy-object-proxy==1.9.0
mccabe==0.7.0
more-itertools==10.1.0
multidict==6.0.4
mypy-extensions==1.0.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pylint==2.17.5
python-dotenv==1.0.0
pytz==2023.3
SQLAlchemy==2.0.20
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.12.1
typing_extensions==4.7.1
wrapt==1.15.0
yarl==1.9.2
